requirements.txt
flask_mongoengine==1.0.0
Flask_ReCaptcha==0.4.2
requests_oauthlib==1.3.0
Flask==1.1.2
mongoengine==0.22.1
qrcode==6.1
Flask_RESTful==0.3.8
requests==2.25.1
APScheduler==3.7.0
omegaconf==2.0.6
Flask_SocketIO==5.0.1
pycryptodome==3.10.1
